pyproject.toml
[project]
name = "project-reporter"
version = "0.1.0"
description = "Multi-project documentation hub with AI-powered ingestion and Discord notifications"
requires-python = ">=3.11"
dependencies = [
    "mkdocs-material>=9.5",
    "mkdocs-awesome-pages-plugin>=2.9",
    "anthropic>=0.40",
    "discord.py>=2.3",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "voyageai>=0.3",
    "numpy>=1.26",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
